+++ Data/QuestChains/_Index.lua
@@ -12,15 +12,16 @@ DBmod:RegisterExpansion(ns.EXP_MIDNIGHT, {
 })
 
 ns.CAT = {
-    CAMPAIGN         = 1100,
-    EVERSONG_WOODS   = 1101,
-    ZULAMAN          = 1102,
-    HARANDAR         = 1103,
-    VOIDSTORM        = 1104,
-    ARATOR           = 1105,
-    WAR_LIGHT_SHADOW = 1106,
-    REVELATIONS      = 1107,
-    VOID_ACROPOLIS   = 1108,
+    CAMPAIGN          = 1100,
+    EVERSONG_WOODS    = 1101,
+    ZULAMAN           = 1102,
+    HARANDAR          = 1103,
+    VOIDSTORM         = 1104,
+    ARATOR            = 1105,
+    WAR_LIGHT_SHADOW  = 1106,
+    REVELATIONS       = 1107,
+    VOID_ACROPOLIS    = 1108,
+    SUNSTRIDER_OMNIUM = 1109,
 }
 
 DBmod:RegisterCategory(ns.CAT.CAMPAIGN,         { expansion = ns.EXP_MIDNIGHT, name = "Midnight Campaign", mapIDs = {}, campaignID = 270, order = 10 })
@@ -31,4 +32,5 @@ DBmod:RegisterCategory(ns.CAT.HARANDAR,         { expansion = ns.EXP_MIDNIGHT, n
 DBmod:RegisterCategory(ns.CAT.ARATOR,           { expansion = ns.EXP_MIDNIGHT, name = "Arator",         mapIDs = {}, order = 60 })
 DBmod:RegisterCategory(ns.CAT.VOIDSTORM,        { expansion = ns.EXP_MIDNIGHT, name = "Voidstorm",      mapIDs = { 2405 }, order = 70 })
 DBmod:RegisterCategory(ns.CAT.REVELATIONS,      { expansion = ns.EXP_MIDNIGHT, name = "Revelations (12.0.7)", mapIDs = { 2424 }, order = 80 })
+DBmod:RegisterCategory(ns.CAT.SUNSTRIDER_OMNIUM, { expansion = ns.EXP_MIDNIGHT, name = "The Sunstrider Omnium", mapIDs = {}, order = 85 })
 DBmod:RegisterCategory(ns.CAT.VOID_ACROPOLIS,   { expansion = ns.EXP_MIDNIGHT, name = "Void Acropolis", mapIDs = { 2599 }, order = 90 })